experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "MedSigLIP fine-tuning experiments for Visilant ophthalmology"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.51.0",
    "accelerate>=1.6.0",
    "peft>=0.15.0",
    "datasets>=3.5.0",
    "evaluate>=0.4.0",
    "sentencepiece>=0.2.0",
    "scikit-learn>=1.8.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "pillow>=12.0.0",
    "wandb>=0.19.0",
    "pyyaml>=6.0",
    "matplotlib>=3.10.0",
    "pytest>=8.0",
]

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu128"]
